Test in a VM on Dawrin
